<==== Mercurial data ====>
Branch: cloud_19.1

+++ vms/distribution/deb/mediaserver/build_distribution.sh
@@ -14,7 +14,7 @@ stripIfNeeded() # dir
 {
     local -r DIR="$1" && shift
 
-    if [[ "$BUILD_CONFIG" == "Release" && "$ARCH" != "aarch64" ]]
+    if [[ $BUILD_CONFIG == "Release" && $ARCH != "arm64" && $TARGET_DEVICE != "linux_arm32" ]]
     then
         local FILE
         for FILE in $(find "$DIR" -type f)
@@ -38,7 +38,6 @@ copyLibs()
     echo ""
     echo "Copying libs"
 
-    mkdir -p "$STAGE_LIB"
     local LIB
     local LIB_BASENAME
     local BLACKLIST_ITEM
@@ -60,14 +59,21 @@ copyLibs()
         'libvms_gateway_core*'
     )
 
-    for LIB in "$BUILD_DIR/lib"/*.so*
+    for LIB in "$BUILD_DIR/lib"/*.so* "$BUILD_DIR/lib"/ffmpeg-*/*.so*
     do
+        # Skip non-resolved globs.
+        [[ -r $LIB ]] || continue
+
+        local RELATIVE_PATH=$(dirname ${LIB#"$BUILD_DIR/lib/"})
+
         LIB_BASENAME=$(basename "$LIB")
 
         SKIP_LIBRARY=0
 
-        for BLACKLIST_ITEM in "${LIB_BLACKLIST[@]}"; do
-            if [[ $LIB_BASENAME == $BLACKLIST_ITEM ]]; then
+        for BLACKLIST_ITEM in "${LIB_BLACKLIST[@]}"
+        do
+            if [[ $LIB_BASENAME == $BLACKLIST_ITEM ]]
+            then
                 SKIP_LIBRARY=1
                 break
             fi
@@ -75,17 +81,14 @@ copyLibs()
 
         (( $SKIP_LIBRARY == 1 )) && continue
 
-        if [[ "$LIB_BASENAME" == libtegra_video.* ]]; then
-            [[ "$BOX" != "tx1" ]] && continue
-        fi
-
         echo "  Copying $LIB_BASENAME"
-        cp -P "$LIB" "$STAGE_LIB/"
+        mkdir -p "$STAGE_LIB/${RELATIVE_PATH}/"
+        cp -P "$LIB" "$STAGE_LIB/${RELATIVE_PATH}/"
     done
 
     echo "  Copying system libs: ${CPP_RUNTIME_LIBS[@]}"
     distrib_copySystemLibs "$STAGE_LIB" "${CPP_RUNTIME_LIBS[@]}"
-    if [[ "$ARCH" != "arm" ]]
+    if [[ "$ARCH" != "arm" && $TARGET_DEVICE != "linux_arm32" ]]
     then
         echo "Copying libicu"
         distrib_copySystemLibs "$STAGE_LIB" "${ICU_RUNTIME_LIBS[@]}"
@@ -103,8 +106,7 @@ copyMediaserverPlugins()
     distrib_copyMediaserverPlugins "plugins" "$STAGE_MODULE/bin" "${SERVER_PLUGINS[@]}"
     stripIfNeeded "$STAGE_MODULE/bin/plugins"
 
-    distrib_copyMediaserverPlugins "plugins_optional" "$STAGE_MODULE/bin" \
-        "${SERVER_PLUGINS_OPTIONAL[@]}"
+    distrib_copyMediaserverPlugins "plugins_optional" "$STAGE_MODULE/bin" "${SERVER_PLUGINS_OPTIONAL[@]}"
     stripIfNeeded "$STAGE_MODULE/bin/plugins_optional"
 }
 
@@ -114,6 +116,39 @@ copyFestivalVox()
     echo "Copying Festival Vox files"
     mkdir -p "$STAGE_BIN"
     cp -r "$BUILD_DIR/bin/vox" "$STAGE_BIN/"
+}
+
+copyFiles()
+{
+    echo "Copying generic files"
+
+    if [[ -d $FILES_DIR/$TARGET_DEVICE ]] \
+        && (( $(ls -Aq "$FILES_DIR/$TARGET_DEVICE" |wc -w) > 0 ))
+    then
+       cp -r "$FILES_DIR/$TARGET_DEVICE/"* "$STAGE_MODULE/"
+    fi
+}
+
+# [in] STAGE_LIB
+# [in] STAGE_MODULE
+# [in] SOURCE_DIR
+copyTegraSpecificFiles()
+{
+    echo ""
+    echo "Copying Tegra-specific files"
+
+    mkdir -p "$STAGE_LIB"
+
+    echo "  Copying libtegra_video.so"
+    cp -P "$BUILD_DIR/lib/libtegra_video.so" "$STAGE_LIB/"
+
+    local -r TEGRA_VIDEO_SOURCE_DIR="$SOURCE_DIR/artifacts/tx1/tegra_multimedia_api"
+    # Demo neural networks.
+    local -r NVIDIA_MODELS_SOURCE_DIR="$TEGRA_VIDEO_SOURCE_DIR/data/model"
+    local -r NVIDIA_MODELS_INSTALL_PATH="$STAGE_MODULE/nvidia_models"
+
+    mkdir -p "$NVIDIA_MODELS_INSTALL_PATH"
+    cp -f $NVIDIA_MODELS_SOURCE_DIR/* "$NVIDIA_MODELS_INSTALL_PATH"
 }
 
 # [in] STAGE_LIB
@@ -160,7 +195,7 @@ copyBins()
 {
     echo "Copying mediaserver binaries and scripts"
     install -m 755 "$BUILD_DIR/bin/mediaserver" "$STAGE_BIN/mediaserver-bin"
-    if [ "$ENABLE_ROOT_TOOL" = "true" ]
+    if [[ $ENABLE_ROOT_TOOL == "true" ]]
     then
         install -m 750 "$BUILD_DIR/bin/root_tool" "$STAGE_BIN/root-tool-bin"
     fi
@@ -169,6 +204,7 @@ copyBins()
     install -m 644 "$CURRENT_BUILD_DIR/qt.conf" "$STAGE_BIN/"
     install -m 755 "$SCRIPTS_DIR/config_helper.py" "$STAGE_BIN/"
     install -m 755 "$SCRIPTS_DIR/shell_utils.sh" "$STAGE_BIN/"
+    install -m 755 "$SOURCE_DIR/nx_log_viewer.html" "$STAGE_BIN/"
 
     echo "Copying mediaserver startup script"
     install -m 755 "bin/mediaserver" "$STAGE_BIN/"
@@ -182,14 +218,19 @@ copyStartupScripts()
 
     install -m 644 "init/networkoptix-mediaserver.conf" \
         "$STAGE/etc/init/$CUSTOMIZATION-mediaserver.conf"
-    install -m 644 "init/networkoptix-root-tool.conf" \
-        "$STAGE/etc/init/$CUSTOMIZATION-root-tool.conf"
 
     install -m 755 -d "$STAGE/etc/systemd/system"
+
     install -m 644 "systemd/networkoptix-mediaserver.service" \
         "$STAGE/etc/systemd/system/$CUSTOMIZATION-mediaserver.service"
-    install -m 644 "systemd/networkoptix-root-tool.service" \
-        "$STAGE/etc/systemd/system/$CUSTOMIZATION-root-tool.service"
+
+    if [[ $ENABLE_ROOT_TOOL == "true" ]]
+    then
+        install -m 644 "init/networkoptix-root-tool.conf" \
+            "$STAGE/etc/init/$CUSTOMIZATION-root-tool.conf"
+        install -m 644 "systemd/networkoptix-root-tool.service" \
+            "$STAGE/etc/systemd/system/$CUSTOMIZATION-root-tool.service"
+    fi
 }
 
 # [in] STAGE
@@ -242,14 +283,32 @@ createUpdateZip() # file.deb
         cp -r "$DEB" "$ZIP_DIR/"
     done
 
-    local FILE
-    for FILE in "update"/*
-    do
+    local FILE="update/install.sh"
+    if [[ -f $FILE ]]
+    then
         echo "  Copying (configured) $(basename "$FILE")"
         cp -r "$FILE" "$ZIP_DIR/"
-    done
+    fi
 
+    if [[ -d update/$TARGET_DEVICE ]] && (( $(ls -Aq "update/$TARGET_DEVICE" |wc -w) > 0 ))
+    then
+        for FILE in "update/$TARGET_DEVICE/"*
+        do
+            echo "  Copying (configured) $(basename "$FILE")"
+            cp -r "$FILE" "$ZIP_DIR/"
+        done
+    fi
+
+    cp -r "update/update.json" "$ZIP_DIR/update.json"
     distrib_createArchive "$DISTRIBUTION_OUTPUT_DIR/$UPDATE_ZIP" "$ZIP_DIR" zip -r
+
+    if [[ $TARGET_DEVICE == "linux_arm32" ]]
+    then
+        cp -r "update/update.rpi.json" "$ZIP_DIR/update.json"
+        distrib_createArchive "$DISTRIBUTION_OUTPUT_DIR/$UPDATE_ZIP_RPI" "$ZIP_DIR" zip -r
+        cp -r "update/update.bananapi.json" "$ZIP_DIR/update.json"
+        distrib_createArchive "$DISTRIBUTION_OUTPUT_DIR/$UPDATE_ZIP_BANANAPI" "$ZIP_DIR" zip -r
+    fi
 }
 
 # [in] WORK_DIR
@@ -273,6 +332,12 @@ buildDistribution()
     copyQtLibs
     copyMediaserverPlugins
     copyFestivalVox
+    copyFiles
+
+    if [[ $TARGET_DEVICE == 'linux_arm64' ]]
+    then
+        copyTegraSpecificFiles
+    fi
 
     echo "Setting permissions"
     find "$STAGE" -type d -print0 |xargs -0 chmod 755
